chore: rename default ticket to derivative ticket
