+++ apps/trading/components/ticket/derivative/feedback-stop.tsx
