+++ apps/trading/components/ticket/derivative/stop-limit.tsx
