+++ apps/trading/components/ticket/derivative/stop-market.tsx
