+++ apps/trading/components/ticket/derivative/ticket.tsx
