+++ apps/trading/components/ticket/derivative/utils.ts
