+++ apps/trading/components/ticket/spot/feedback-stop.tsx
