+++ apps/trading/components/ticket/spot/feedback.tsx
