+++ apps/trading/components/ticket/spot/fields/index.ts
