+++ apps/trading/components/ticket/spot/fields/notional.tsx
