+++ apps/trading/components/ticket/spot/fields/size-slider.tsx
