+++ apps/trading/components/ticket/spot/fields/size.tsx
