+++ apps/trading/components/ticket/spot/fields/stop-size.tsx
