+++ apps/trading/components/ticket/spot/market.tsx
