+++ apps/trading/components/ticket/spot/stop-limit.tsx
