+++ apps/trading/components/ticket/spot/stop-market.tsx
