+++ apps/trading/components/ticket/spot/ticket.tsx
