+++ apps/trading/components/ticket/spot/utils.ts
